requirements.txt
gradio
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
numpy==1.26.4
opencv-python
matplotlib
scikit-image
shap==0.44.1
